move thread-panic tests to their own file; test getting the thread name
