+++ tests/pass/concurrency/simple.rs
@@ -45,23 +45,6 @@ fn create_move_out() {
     assert_eq!(result.len(), 6);
 }
 
-fn panic() {
-    let result = thread::spawn(|| panic!("Hello!")).join().unwrap_err();
-    let msg = result.downcast_ref::<&'static str>().unwrap();
-    assert_eq!(*msg, "Hello!");
-}
-
-fn panic_named() {
-    thread::Builder::new()
-        .name("childthread".to_string())
-        .spawn(move || {
-            panic!("Hello, world!");
-        })
-        .unwrap()
-        .join()
-        .unwrap_err();
-}
-
 // This is not a data race!
 fn shared_readonly() {
     use std::sync::Arc;
@@ -89,6 +72,4 @@ fn main() {
     create_move_in();
     create_move_out();
     shared_readonly();
-    panic();
-    panic_named();
 }
