--- tests/pass/concurrency/simple.stderr
@@ -1,5 +1,0 @@
-thread '<unnamed>' panicked at $DIR/simple.rs:LL:CC:
-Hello!
-note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
-thread 'childthread' panicked at $DIR/simple.rs:LL:CC:
-Hello, world!
